Removed marching triangles mesh in favor of clipped marching squares for efficiency in contour finding
